Tests mirror src/: five suites move under tests/domain/ (#195)
Pure git renames + include-depth fix + Makefile glob/rule paths;
telemetry/ and alerts/ already mirrored. Zero expectation changes,
27 binaries green. TESTING.md tree redrawn (it had drifted: alerts/
and test_observer_encoders were missing); FILE-MAP updated.
